refactor: useTetrisGame 훅으로 UI와 데이터를 분리
