+++ tetris/src/pages/play/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useTetrisGame } from './useTetrisGame';
